+++ CHANGELOG.md
@@ -12,6 +12,7 @@ TBD
 #### Enhancements
 
 - Support Xcode 7.3 and Swift 2.2
+- Configurable mask polygon (sides)
 - Add `SystemPageCurlAnimator` to support `SystemPageCurlFromTop` and `SystemPageCurlFromBottom` transition animations
 
 #### Bugfixes